Started to build up the Game Manager. - Added a method to Game_Manager to list all games available to install.
NOTE: I'm still not entirely comfortable with this layout for games yet. The game_wow.php file is essentially static information, so it doesn't really need to be a class I don't think. Yet, at the same time, game.php does make sense to be an object. I may end up altering this so that game_wow.php is a classless file and game.php pulls information from it as it is required.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@355 0a1ec8ee-f233-0410-a8d7-37675f9b97f8