Start of my investigation into the dreck that is the Session class and its high resource usage. Prior to this revision, an anonymous user refreshing the page rapidly resulted in hundreds and hundreds of sessions being created, and 'cleanup' being called often, despite it being one user from one IP address.
This revision attempts to limit that behavior, although this code is showing its age and is as confusing as it is poor-performing.

I've added debugging 'output' throughout the code, but since this class handles a lot of cookies, we cheat the system by outputting these messages via DBAL's query array, to be output later, rather than echo'ing them immediately and breaking setcookie().

NOTE: In order to decide when to perform "dead session" cleanup, this script checks the current time against a config variable called session_cleanup. My value was initially set to 0, resulting in cleanup being called every second, if you refreshed quickly enough. For testing, I set my value to something more sane like 600. This is something we'll need to set in the upgrade script if it's 0 in every installation.

How's that for a commit message?!

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@404 0a1ec8ee-f233-0410-a8d7-37675f9b97f8