Instead of trying to continue working on the current Session and User classes, I decided to rewrite them, because I also wanted to enhance the security of User passwords using salts while at the same time gaining a much better understanding of how the classes (should) work.
This file is a self-contained test of the new methods. It mostly works a *lot* better than before, in my personal testing, however it's not ready to be dropped into the uploads folder for real use; both because it has a few minor quirks, and because it requires major database changes that I haven't added to the schema or upgrade files. More to come.

Committing now before sleep deprivation makes me lose five hours of work.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@407 0a1ec8ee-f233-0410-a8d7-37675f9b97f8