Creating log-parsing code
The direction I'm hoping to take is starting to shape up.
What we do is match the format for a magic code, and store the results in an array. Then, we iterate over the results. We end up with three things:
  - a partial string which matches 
    - all the leading characters before the magic string
    - the magic string itself
    - all trailing characters before the next magic string
  - the offset of the match
  - the magic string's type (name, guild, etc.)

The plan is that from here, one can create a switch statement which uses all of the above information to extract the data from the log entry, according to the rules associated with extracting data for each magic string type. 
In other words, you can create functions that will deal with each magic string type separately, and these functions can convert the magic string into a regular expression or otherwise to retrieve the data.

The downside so far is that in terms of relative complexity and operation time, this could end up being inadequate for actually parsing logs in real-time. I'll have to do some time tests once I get it functioning correctly.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@463 0a1ec8ee-f233-0410-a8d7-37675f9b97f8