Log parsing Creating log-parsing code
- Removed the flag for PREG_OFFSET_CAPTURE, because funnily enough, string positions inside the regular expression string don't mean anything :P
- Added the ability to detect optional groupings within the parse string
- Added some comments on how the expression works, and what it will return.

NOTE:
Now that I think about it, the stuff to capture the optional component stuff may not be necessary, depending on how the specific functions work. If the sub method were to discover an optional component, it could simply pass the parse_log_entry method the substring for both the log entry to match, and the partial parse string.
In other words, you could use recursion to grab the data for optional components, and merge that data into the root data when the recursion returns.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@464 0a1ec8ee-f233-0410-a8d7-37675f9b97f8