Creating log-parsing code
- Added the doing methods for the parsing process
- Added fuckloads of echo's, so when you run tests/parse.php you can see what's happening.
- The code is documented well enough I hope - go read it.

NOTE:
Actually assigning the discovered values to the return array hasn't happened yet. It follows that optional groupings haven't been dealt with yet either. However, I know how i'm going to do it. 

What it will entail is each of the new methods returning an array of the form ($datatype => $value). When a method receives the results from a method it just called, it will have to merge the results into a larger array that it will eventually return itself. 

In this way, optional groupings can be dealt with by simply isolating that component of the parse string / result set, and either adding that to the returning array in the current method, or calling another method and merging the results once they're returned.

Hope that makes a little bit of sense.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@465 0a1ec8ee-f233-0410-a8d7-37675f9b97f8