Log Parsing
As-is, the logs seem to be parsing well. I know there's room for improvement in efficiency in places, but for now it seems to be doing well enough. Granted the guild, zone and date/time need to be parsed and returned still, but the biggest time increase there will be the date, which I hope will not impact all too much on performance.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@467 0a1ec8ee-f233-0410-a8d7-37675f9b97f8