Log Parsing
- Re-worked methods in order to allow for recursive calls.
- Added support for parsing optional groups via recursion.

Unfortunately, the way I've written this seems to increase the time taken for a full log parse from around 0.027 up to 0.062. There are ways of optimizing this - namely, re-working the methodology so that all of the parse strings are calculated and split before the values are searched for using them. From the results of the testing here, if recursion isn't used things'll be a bit faster. Still, it's not all that bad when you think about it.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@470 0a1ec8ee-f233-0410-a8d7-37675f9b97f8