Database Structure, Game support (IMPORTANT!)
NOTE: Some pages will not work as a result of this update. That's to be expected - I have to run through all the pages and find which ones broke and need updating to use these changes. This was just a logical save point.

Added new armor_types and class_armor tables into the mix. This requires a few things to be updated to make use of the new structure.

- Updated installation procedure to throw data into the new tables instead of the old class table.
- Added some TODOs to the upgrade-1.4.0_01 file regarding the new table changes.


In addition, each game-specific table now has a xxx_key column. This column is meant to store an alpha-numeric (and possibly underscore) string which can be used to retrieve a language-specific spelling for a named thing in those tables. While the functions haven't been implemented yet, the flow will go something like this:
 - The game-specific language file is included on session startup.
 - Whenever the data from these tables is queried through the game manager, the method will get the xxx_key value, and check to see if an entry for that key exists in the global language array.
   eg: $localized_name = isset($lang['class_' + $db_key]) ? $lang['class_' + $db_key] : $db_name;
 - There should be a method to do the conversions noted above so that it can be done consistently by any part of EQdkp that may wish to call on this functionality (ie: plugins).



git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@478 0a1ec8ee-f233-0410-a8d7-37675f9b97f8