Cleanup of old games (IMPORTANT!)
- Removing these old, defunct files from the project. 

These are being removed now, because it makes things easier in terms of finding all the locations in EQdkp affected by the recent database changes by simply searching for class_armor_type or another old column name.

Of the games listed here, Everquest 2, Vanguard and Dark age of Camelot have not been re-written for the new game file format. This can be a point of reference for those games when the time comes, but for now they're gone.

Also, the game.php file was never used. It's an old relic left over from when I was going to do game management in a more object-oriented way.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@483 0a1ec8ee-f233-0410-a8d7-37675f9b97f8