Installation, Input cleaning
This is always a fun game to play =P
- Added extra restrictions to the database table prefix name.
- Updated the error message when an invalid table name is given.
- Added a function to properly format the data for the config.php file when it is written to prevent invalid PHP code from being written to the file.

NOTE: Under the MySQL identifier schema, characters such as apostrophe are valid in table names. However, when used in conjunction with PHP it becomes rather irritating to correctly support them. So, for now, they're invalid characters. Feel free to send me death threats if this affects you badly.

git-svn-id: https://eqdkp.googlecode.com/svn/branches/inputsec/@547 0a1ec8ee-f233-0410-a8d7-37675f9b97f8